⚡ Bolt: Replace array reverse clone with findLast utility (#20)
* perf: replace array reverse clone with findLast utility

Replaces instances of `[...arr].reverse().find()` and `arr.slice().reverse().find()` with a custom `findLast` utility. This avoids O(n) memory allocation and O(n) time overhead when searching for an element backwards in arrays, providing a measurable performance improvement for finding latest tool calls or session messages, particularly in performance sensitive components like `Chat` and `OrchestratorPanel` when dealing with longer chat histories.


* perf: replace array reverse clone with findLast utility

Replaces instances of `[...arr].reverse().find()` and `arr.slice().reverse().find()` with a custom `findLast` utility. This avoids O(n) memory allocation and O(n) time overhead when searching for an element backwards in arrays, providing a measurable performance improvement for finding latest tool calls or session messages, particularly in performance sensitive components like `Chat` and `OrchestratorPanel` when dealing with longer chat histories.

Also resolves a backend lint failure related to `too_many_arguments` in `tandem-enterprise-contract` by adding `#[allow(clippy::too_many_arguments)]`.


* perf: replace array reverse clone with findLast utility

Replaces instances of `[...arr].reverse().find()` and `arr.slice().reverse().find()` with a custom `findLast` utility. This avoids O(n) memory allocation and O(n) time overhead when searching for an element backwards in arrays, providing a measurable performance improvement for finding latest tool calls or session messages, particularly in performance sensitive components like `Chat` and `OrchestratorPanel` when dealing with longer chat histories.

Also resolves a backend lint failure related to `too_many_arguments` in `tandem-enterprise-contract` by adding `#[allow(clippy::too_many_arguments)]`, fixes clippy warnings in `tandem-tools`, and fixes an unnecessary cast (`u64` -> `u64`) in `tandem-governance-engine`.


* perf: replace array reverse clone with findLast utility

Replaces instances of `[...arr].reverse().find()` and `arr.slice().reverse().find()` with a custom `findLast` utility. This avoids O(n) memory allocation and O(n) time overhead when searching for an element backwards in arrays, providing a measurable performance improvement for finding latest tool calls or session messages, particularly in performance sensitive components like `Chat` and `OrchestratorPanel` when dealing with longer chat histories.

Also resolves a backend lint failures across the rust crates, including:
- resolving too_many_arguments in orchestrator/engine.rs and tandem-enterprise-contract
- collapsing `if` blocks into outer `match` patterns inside `orchestrator/engine.rs` and `ralph/service.rs`
- using `sort_by_key` instead of `sort_by` for more efficient sorting
- cleaning up unnecessary borrows and unused assignments in src-tauri


---------

+++ crates/tandem-governance-engine/src/lib.rs
@@ -517,7 +517,7 @@ impl GovernancePolicyEngine for DefaultGovernanceEngine {
             });
         }
         if limits.health_guardrail_stop_threshold > 0
-            && input.guardrail_stop_count >= limits.health_guardrail_stop_threshold as u64
+            && input.guardrail_stop_count >= limits.health_guardrail_stop_threshold
         {
             findings.push(AutomationLifecycleFinding {
                 finding_id: format!("finding-{}", Uuid::new_v4().simple()),
